chatbot/requirements.txt
# RAG Chatbot Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# LLM Providers
openai>=1.12.0
tiktoken>=0.6.0
google-genai>=1.0.0

# Vector Database (Qdrant)
qdrant-client>=1.7.0

# PostgreSQL (Neon)
asyncpg>=0.29.0

# Configuration & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Markdown Processing
python-frontmatter>=1.1.0

# Development
python-dotenv>=1.0.0

# Optional: Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
